Add required code after file renaming
